requirements.txt
requests==2.31.0
pandas==2.0.2
seaborn==0.12.2
tqdm==4.64.0
matplotlib==3.7.1
gitpython==3.1.31
